more helper functions and components refactored
