+++ src/Components/allocations/AllocationItemEditable.tsx
@@ -1,7 +1,7 @@
 import { Button, TextField } from "@mui/material";
-import { Dispatch, SetStateAction, useState } from "react";
-import { updateDataInFirebase } from "../../../dataBaseUtils/readWrite";
-import { allocationType, editType } from "../../../types";
+import { ChangeEvent, Dispatch, SetStateAction, useState } from "react";
+import { updateDataInFirebase } from "../../dataBaseUtils/readWrite";
+import { allocationType, editType } from "../../types";
 
 type Props = {
   allocation: allocationType;
@@ -12,7 +12,7 @@ type Props = {
 const AllocationItemEditable = (props: Props) => {
   const { allocation, setEdit, setReRender } = props;
   const [allocationName, setAllocationName] = useState<string>(allocation?.name || "");
-  const handleAllocationNameChange = (event: any) => {
+  const handleAllocationNameChange = (event: ChangeEvent<HTMLInputElement>) => {
     setAllocationName(event.target.value)
   }
   const handleAllocationNameSave = () => {
